backend/requirements.txt
django>=5.2
djangorestframework>=3.16
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
dj-database-url>=2.2
python-dotenv>=1.0
psycopg2-binary>=2.9
pypdf>=5.0
pymupdf>=1.24.0
python-docx>=1.1
Pillow>=10.0
scikit-learn>=1.5
openai>=1.40.0
